Class

+++ Class/Class.Port/Read.cs
@@ -758,13 +758,13 @@ public class Read : Any
         return a;
     }
 
-    protected virtual Array ExecuteExportArray(int row, int lineCount)
+    protected virtual Array ExecuteExportArray(long row, long lineCount)
     {
-        int count;
+        long count;
         count = lineCount;
         Array array;
         array = this.Operate.ExecuteArray(count);
-        int i;
+        long i;
         i = 0;
         while (i < count)
         {
